+++ Source/OS9exec_core/filescsi.h
@@ -55,13 +55,13 @@
 
 /* Low level SCSI interface */
 
-os9err Set_SSize   ( scsi_dev* scsi,                 ulong  sctSize );
-os9err Get_SSize   ( scsi_dev* scsi,                 ulong *sctSize );
-os9err ReadCapacity( scsi_dev* scsi, ulong *totScts, ulong *sctSize );
-os9err Get_DSize   ( scsi_dev* scsi, ulong *totScts );
+os9err Set_SSize   ( scsi_dev* scsi,                    uint32_t  sctSize );
+os9err Get_SSize   ( scsi_dev* scsi,                    uint32_t *sctSize );
+os9err ReadCapacity( scsi_dev* scsi, uint32_t *totScts, uint32_t *sctSize );
+os9err Get_DSize   ( scsi_dev* scsi, uint32_t *totScts );
 
-os9err ReadFromSCSI( scsi_dev* scsi, ulong sectorNr, ulong nSectors, ulong len, byte* buffer );
-os9err WriteToSCSI ( scsi_dev* scsi, ulong sectorNr, ulong nSectors, ulong len, byte* buffer );
+os9err ReadFromSCSI( scsi_dev* scsi, uint32_t sectorNr, uint32_t nSectors, uint32_t len, byte* buffer );
+os9err WriteToSCSI ( scsi_dev* scsi, uint32_t sectorNr, uint32_t nSectors, uint32_t len, byte* buffer );
 
 #ifdef windows32
 void scsiadaptor_help (void);
